+++ CHANGELOG.md
@@ -22,6 +22,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Turn on checkstyle JavaDocs module.
 - Add updates to the protocol, like new `ControlMessage`.
 
+## [1.6.3](https://github.com/appulse-projects/encon-java/releases/tag/1.6.3) - 2018-09-11
+
+### Changed
+
+- Removed multi-handlers `netty` pipeline, use single one instead;
+- Cache `atom` and `pid` values;
+- Improve benchmarks.
+
 ## [1.6.2](https://github.com/appulse-projects/encon-java/releases/tag/1.6.2) - 2018-09-06
 
 ### Added